feature: Add alert for expiration tests fails
This alert checks the result of domain and certificate expiration checks. We are 
using 24h intervals in Prometheus query as these tests are run only once a day 
and can fail only once a day. The default alert for failed checks will not catch 
it. 

The default alert has the times adjusted a little, as there could be a case 
where a 1m interval would fit exactly between the test runs and not alert about 
failed checks.

Touches: #17.